+++ src/main/java/de/chrgroth/jsonstore/json/flexjson/FlexjsonService.java
@@ -1,7 +1,6 @@
 package de.chrgroth.jsonstore.json.flexjson;
 
 import java.lang.reflect.Method;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -25,8 +24,6 @@ import flexjson.JSONSerializer;
 import flexjson.JSONTokener;
 import flexjson.JsonNumber;
 import flexjson.ObjectBinder;
-
-// TODO move some fromJson migration logic to abstract class?
 
 public final class FlexjsonService implements JsonService {
     private static final Logger LOG = LoggerFactory.getLogger(FlexjsonService.class);
@@ -220,8 +217,6 @@ public final class FlexjsonService implements JsonService {
         }
     }
 
-    // TODO drop scenario of backward compatibility without metadata being used
-
     @Override
     public boolean fromJson(JsonStoreMetadata<?, ?> metadata, Map<Integer, VersionMigrationHandler> migrationHandlers, String json) {
 
@@ -235,59 +230,59 @@ public final class FlexjsonService implements JsonService {
         Object genericStructureRaw = new JSONTokener(json).nextValue();
         stopwatch.stop();
         LOG.info(metadata.getUid() + ": raw parsing from json took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
-        if (genericStructureRaw == null) {
+        if (!(genericStructureRaw instanceof Map)) {
             return false;
         }
 
-        // determine current data situation
-        boolean isMap = genericStructureRaw instanceof Map;
+        // ensure metadata wrapper
         @SuppressWarnings("unchecked")
-        boolean isMetadataAvailable = isMap && JsonStoreMetadata.class.getName().equals(((Map<String, Object>) genericStructureRaw).get(JSON_FIELD_CLASS));
-        @SuppressWarnings("unchecked")
-        boolean isSingleton = isMetadataAvailable ? (boolean) ((Map<String, Object>) genericStructureRaw).get(JSON_FIELD_SINGLETON) : isMap;
-
-        // determine generic payload
-        @SuppressWarnings("unchecked")
-        Object genericStructurePayload = isMetadataAvailable ? ((Map<String, Object>) genericStructureRaw).get(JSON_FIELD_PAYLOAD) : genericStructureRaw;
-        if (genericStructurePayload == null) {
+        Map<String, Object> metadataRaw = (Map<String, Object>) genericStructureRaw;
+        boolean isMetadataAvailable = JsonStoreMetadata.class.getName().equals(metadataRaw.get(JSON_FIELD_CLASS));
+        if (!isMetadataAvailable) {
+            LOG.error(metadata.getUid() + ": json invalid, no metadata wrapper detected.");
             return false;
         }
-
-        // determine version information
-        @SuppressWarnings("unchecked")
-        Object topLevelTypeVersionRaw = isMetadataAvailable ? ((Map<String, Object>) genericStructureRaw).get(JSON_FIELD_PAYLOAD_TYPE_VERSION) : null;
-        Integer topLevelTypeVersion = topLevelTypeVersionRaw != null ? ((JsonNumber) topLevelTypeVersionRaw).toInteger() : 0;
-        Integer payloadTypeVersion = metadata.getPayloadTypeVersion();
 
         // migrate payload data
-        boolean migrated = migrateVersions(metadata, migrationHandlers, isSingleton, topLevelTypeVersion, payloadTypeVersion, genericStructurePayload);
+        boolean migrated = migrateVersions(metadata, migrationHandlers, metadataRaw);
 
         // process deserialization to payload object instances
-        jsonDeserialization(metadata, genericStructureRaw, isMetadataAvailable, genericStructurePayload);
+        jsonDeserialization(metadata, metadataRaw);
 
         return migrated;
     }
 
     @SuppressWarnings("unchecked")
-    private boolean migrateVersions(JsonStoreMetadata<?, ?> metadata, Map<Integer, VersionMigrationHandler> migrationHandlers, boolean isSingleton, Integer topLevelTypeVersion,
-            Integer payloadTypeVersion, Object genericStructurePayload) {
+    private boolean migrateVersions(JsonStoreMetadata<?, ?> metadata, Map<Integer, VersionMigrationHandler> migrationHandlers, Map<String, Object> metadataRaw) {
+
+        // determine payload
+        Object rawPayload = metadataRaw.get(JSON_FIELD_PAYLOAD);
+        if (rawPayload == null) {
+            return false;
+        }
+
+        // determine source information
+        boolean isSingleton = (boolean) metadataRaw.get(JSON_FIELD_SINGLETON);
+        Object sourceTypeVersionRaw = metadataRaw.get(JSON_FIELD_PAYLOAD_TYPE_VERSION);
+        Integer sourceTypeVersion = sourceTypeVersionRaw != null ? ((JsonNumber) sourceTypeVersionRaw).toInteger() : 0;
 
         // compare version information
-        if (topLevelTypeVersion == null || payloadTypeVersion == null) {
+        Integer targetTypeVersion = metadata.getPayloadTypeVersion();
+        if (sourceTypeVersion == null || targetTypeVersion == null) {
             return false;
         }
 
         // abort on newer version than available as code
-        if (topLevelTypeVersion > payloadTypeVersion) {
-            throw new JsonStoreException("loaded version is newer than specified version in code: " + topLevelTypeVersion + " > " + payloadTypeVersion + "!!");
+        if (sourceTypeVersion > targetTypeVersion) {
+            throw new JsonStoreException("loaded version is newer than specified version in code: " + sourceTypeVersion + " > " + targetTypeVersion + "!!");
         }
 
         // run all available version migrators
         boolean migrated = false;
-        if (topLevelTypeVersion < payloadTypeVersion) {
+        if (sourceTypeVersion < targetTypeVersion) {
 
             // update per version
-            for (int i = topLevelTypeVersion; i <= payloadTypeVersion; i++) {
+            for (int i = sourceTypeVersion; i <= targetTypeVersion; i++) {
 
                 // check for migration handler
                 VersionMigrationHandler migrationHandler = migrationHandlers.get(i);
@@ -300,9 +295,9 @@ public final class FlexjsonService implements JsonService {
                 try {
                     Stopwatch stopwatch = Stopwatch.createStarted();
                     if (isSingleton) {
-                        migrationHandler.migrate((Map<String, Object>) genericStructurePayload);
+                        migrationHandler.migrate((Map<String, Object>) rawPayload);
                     } else {
-                        for (Object genericStructurePayloadItem : (List<Object>) genericStructurePayload) {
+                        for (Object genericStructurePayloadItem : (List<Object>) rawPayload) {
                             migrationHandler.migrate((Map<String, Object>) genericStructurePayloadItem);
                         }
                     }
@@ -320,7 +315,7 @@ public final class FlexjsonService implements JsonService {
     }
 
     @SuppressWarnings("unchecked")
-    private <T, P> void jsonDeserialization(JsonStoreMetadata<T, P> metadata, Object genericStructureRaw, boolean isMetadataAvailable, Object genericStructurePayload) {
+    private <T, P> void jsonDeserialization(JsonStoreMetadata<T, P> metadata, Map<String, Object> metadataRaw) {
 
         // proceed with deserialization
         try {
@@ -332,21 +327,10 @@ public final class FlexjsonService implements JsonService {
             method.setAccessible(true);
             ObjectBinder binder = (ObjectBinder) method.invoke(deserializer);
 
-            if (isMetadataAvailable) {
+            // metadata deserialization
+            JsonStoreMetadata<T, P> oldMetadata = (JsonStoreMetadata<T, P>) binder.bind(metadataRaw);
+            metadata.setPayload(oldMetadata.getPayload());
 
-                // full metadata deserialization
-                JsonStoreMetadata<T, P> oldMetadata = (JsonStoreMetadata<T, P>) binder.bind(genericStructureRaw);
-                metadata.setPayload(oldMetadata.getPayload());
-            } else {
-
-                // proceed payload deserialization
-                metadata.setPayload((P) binder.bind(genericStructurePayload));
-
-                // update metadata timestamps
-                Date now = new Date();
-                metadata.setCreated(now);
-                metadata.setModified(now);
-            }
             stopwatch.stop();
             LOG.info(metadata.getUid() + ": deserializing from json took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
         } catch (Exception e) {