+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 
 0.9.0 (in progress)
 -------------------
+- BREAKING: removed backwards compatibility to load stores without metadata (introduced in version 0.5.0)
 - convenience method to load all stores in case auto save mode is disabled
 - added optional payload class qualifier to be able to handle multiple stores using the same payload class
 - created interface de.chrgroth.jsonstore.json.JsonService regarding all JSON operations and a default implementation for currently used flexjson library